debug add get object meta from PopulateObjectStatus in core_worker.cc to ProcessMessage store.cc

+++ src/ray/object_manager/plasma/protocol.cc
@@ -688,7 +688,7 @@ Status SendMetaReply(const std::shared_ptr<Client> &client, unsigned long long &
   return PlasmaSend(client, MessageType::PlasmaGetMetaReply, &fbb, message);
 }
 
-Status ReadMetaReply(uint8_t *data, size_t size, unsigned long &address, int64_t &object_size, int &device_num) {
+Status ReadMetaReply(uint8_t *data, size_t size, unsigned long long &address, int64_t &object_size, int &device_num) {
   RAY_DCHECK(data);
   auto message = flatbuffers::GetRoot<fb::PlasmaGetMetaReply>(data);
   RAY_DCHECK(VerifyFlatbuffer(message, data, size));
